Move search component out of filter categories
