+++ portal/src/components/InfoBlock/InfoBlock.js
@@ -1,5 +1,5 @@
 import { mapGetters } from 'vuex'
-import Search from '~/components/FilterCategories/Search'
+import Search from '~/components/Search'
 import BreadCrumbs from '~/components/BreadCrumbs'
 import Materials from '~/components/Materials'
