+++ portal/src/components/Search/Search.component.less
@@ -1,4 +1,4 @@
-@import "./../../../variables";
+@import "../../variables";
 .search {
   position: relative;
   z-index: 10;
