+++ portal/src/components/Search/autosuggest.less
@@ -1,4 +1,4 @@
-@import './../../../variables';
+@import '../../variables';
 
 .search {
   #autosuggest {
