+++ portal/src/pages/theme.vue
@@ -93,7 +93,7 @@
 <script>
 import { mapGetters } from 'vuex'
 import _ from 'lodash'
-import Search from '~/components/FilterCategories/Search'
+import Search from '~/components/Search'
 import PopularList from '~/components/Communities/PopularList'
 import Materials from '~/components/Materials'
 import Disciplines from '~/components/Disciplines'
